refactor: move epicenter link utilities from filesystem to workspace package
The epicenter:// URI scheme references workspace concepts (workspace ID,
table name, entity ID)—not filesystem concepts. Moving the link utilities
to @epicenter/workspace eliminates the cross-package import in the
markdown materializer and puts the code where it belongs.

- packages/filesystem/src/links.ts → packages/workspace/src/links.ts
- Remove link re-exports from @epicenter/filesystem barrel
- Consumers import from @epicenter/workspace instead
- Materializer uses relative import (same package now)
